restructure palavyr configuration directory
