+++ frontend/src/dashboard/content/responseConfiguration/conversation/Contracts.ts
