--- frontend/src/dashboard/content/responseConfiguration/conversation/convoDataStructure/nodeOptionals/ShowResponseInPdf.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import { NodeCheckBox } from "../../nodes/nodeInterface/NodeCheckBox";
-import { NodeOptionalProps } from "../Contracts";
-
-export const ShowResponseInPdf = ({ node }: NodeOptionalProps) => {
-    const shouldShow = () => {
-        const nodeTypesThatDoNotProvideFeedback = ["ProvideInfo"];
-        return !node.isTerminal && !nodeTypesThatDoNotProvideFeedback.includes(node.nodeType) && !node.isImageNode;
-    };
-    const onChange = (event: { target: { checked: boolean } }) => {
-        const checked = event.target.checked;
-        node.shouldPresentResponse = checked;
-        node.setTreeWithHistory(node.palavyrLinkedList);
-    };
-
-    return shouldShow() ? <NodeCheckBox label="Show response in PDF" checked={node.shouldPresentResponse} onChange={onChange} /> : <></>;
-};
