+++ frontend/src/dashboard/content/responseConfiguration/conversation/node/imageNode/CustomImage.tsx
