+++ frontend/src/dashboard/content/responseConfiguration/conversation/nodeInterfaceStyles.ts
