--- frontend/src/dashboard/content/responseConfiguration/conversation/treeLines/LineTypes.ts
@@ -1,10 +1,0 @@
-export type LineStyles = "solid" | "-moz-initial" | "inherit" | "initial" | "revert" | "unset" | "dashed" | "dotted" | "double" | "groove" | "hidden" | "inset" | "none" | "outset" | "ridge" | undefined;
-
-export type Anchor = "top" | "left" | "middle" | "center" | "bottom" | "right";
-export type Selector = string;
-
-export type Milliseconds = string;
-export type ParsedAnchor = {
-    x: number;
-    y: number;
-};
